remove todos, WIP fixing hibernate

+++ src/main/java/fr/polytech/rlcalm/controllers/MatchScoreEditController.java
@@ -6,6 +6,7 @@ import fr.polytech.rlcalm.exception.ServiceException;
 import fr.polytech.rlcalm.form.ParticipationUpdateForm;
 import fr.polytech.rlcalm.initializer.ControllerInitializer;
 import fr.polytech.rlcalm.service.MatchService;
+import fr.polytech.rlcalm.utils.FormUtils;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -35,17 +36,17 @@ public class MatchScoreEditController extends HttpServlet {//TODO auth filter
             try {
                 long id = Long.parseLong(parameter);
                 Match match = matchService.getMatch(id);
-                if (match == null) {
-                    //TODO redirect
-                } else {
+                if (match != null) {
                     req.setAttribute("match", match);
                     req.setAttribute("participations1", matchService.getParticipationsOfPlayer1OnMatch(match));
                     req.setAttribute("participations2", matchService.getParticipationsOfPlayer2OnMatch(match));
                     forwardToMatchEdit(req, resp);
+                    return;
                 }
             } catch (NumberFormatException e) {
-                //TODO
+                //redirect
             }
+            resp.sendRedirect("matches");
         } else {
             //creation
             forwardToMatchEdit(req, resp);
@@ -53,44 +54,39 @@ public class MatchScoreEditController extends HttpServlet {//TODO auth filter
     }
 
     private void forwardToMatchEdit(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        forwardToMatchEdit(req, resp, null);
-    }
-
-    private void forwardToMatchEdit(HttpServletRequest req, HttpServletResponse resp, String requestParams) throws ServletException, IOException {
-        //TODO besoin de l'arrayList ?
-        getServletContext().getRequestDispatcher("/match_score_edit.jsp" + (requestParams == null ? "" : requestParams)).forward(req, resp);
+        getServletContext().getRequestDispatcher("/match_score_edit.jsp").forward(req, resp);
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String idParameter = req.getParameter("matchId");
         String action = req.getParameter("action");
         if (action != null) {
             try {
-                long id = Long.parseLong(idParameter);//TODO formUtils ?
-                Match match = matchService.getMatch(id);//TODO error
-                switch (action) {
-                    case "delete": {
-                        matchService.removeScoring(match);
-                        resp.sendRedirect("matches");
-                        break;
-                    }
-                    case "update": {
-                        try {
-                            matchService.updateParticipation(match, ParticipationUpdateForm.formRequest(req));
+                long id = FormUtils.getRequiredLong(req.getParameter("matchId"), null);
+                Match match = matchService.getMatch(id);
+                if (match != null) {
+                    switch (action) {
+                        case "delete": {
+                            matchService.removeScoring(match);
                             resp.sendRedirect("matches");
-                        } catch (InvalidFormException | ServiceException e) {
-                            req.setAttribute("error", e.getMessage());
-                            forwardToMatchEdit(req, resp);
+                            return;
                         }
-                        break;
+                        case "update": {
+                            try {
+                                matchService.updateParticipation(match, ParticipationUpdateForm.formRequest(req));
+                                resp.sendRedirect("matches");
+                            } catch (InvalidFormException | ServiceException e) {
+                                req.setAttribute("error", e.getMessage());
+                                forwardToMatchEdit(req, resp);
+                            }
+                            return;
+                        }
                     }
                 }
             } catch (NumberFormatException e) {
-                //TODO error
+                //redirect
             }
         }
-        //TODO redirect ?
+        resp.sendRedirect("matches");
     }
-
 }
